thermometer: Always write CCC value when connecting
This patch ensures that CCC values for Intermediate Temperature and
Termperature Measurement are always written when connecting to device.
This is to i.e. disable notifications and/or indications in case they
are already enabled (reconnection scenario) but we don't have watcher
registered so it's pointless for remote to send us data.