core: Fix manager_cleanup order
This patch fixes the order in which things are cleaned up in
manager_cleanup. The profiles need to be cleaned up first (since they
are higher level constructs than adapters and devices) and we cannot
remove the adapter from the manager list before calling adapter_remove
since functions within adapter_remove may remove service records which
in turn relies on having the adapter present in the manager list.