audio: Auto connect AVRCP in case of A2DP source is connected
This enables the same logic used for A2DP sink, so it attempt to
connect AVRCP if remote device support it.

Note this is necessary for some devices e.g. WP7 that don't connect
even when being the initiator because it is not mandatory for target
to do so even though it is recommended to avoid collisions while trying
to estabilish the connection.