build-sys: Add missing force arg to symlink creation
If we don't force the symlink creation the build fails when the link
already exists. This happens if the rule to create the link is triggered
again because the file it points to changed. This is the case when doing
incremental builds and we modified the file.

$ make
$ touch lib/cmtp.h
$ make
  GEN      lib/bluetooth/cmtp.h
ln: failed to create symbolic link ‘lib/bluetooth/cmtp.h’: File exists
make: *** [lib/bluetooth/cmtp.h] Error 1
make: *** Waiting for unfinished jobs....