core: Remove stupid adapter path string duplication
Duplicating the path string before calling g_dbus_unregister_interface
is silly. If this is really needed, then there is a bug in the way how
the D-Bus interface unregistration is handled.

If this causes a bug to show up, then lets fix it at the root of it and
not keep working around it.