core: Use adapter->system_name for external configured names
Plugins and other entities can configure the system name. This is different
from the alias set for an adapter and can also be different from the actual
name set by the controller.

This is needed to track the difference between the current controller name,
the configured system name and an user configured alias.