build: Fix --disable-optimization configure option
On commit cc9e4e7cae0379864ea06038d92bf7ecc192bba7, this flag was
mistakenly replaced with the behavior of the old --enable-fortify
option.

This patch restores the "-O0" flag when --disable-optimization is used.

Unfortunately, this is not enough to disable build optimization. By
default, autoconf adds -O2 to CFLAGS if the compiler is GCC. AM_CFLAGS
(where -O0 is added with --disable-optimization) is passed as argument
to GCC before autoconf CFLAGS, so it is not possible to override the
default -O2. One solution is to use:

CFLAGS= ./configure --disable-optimization

i.e. remove -O2 from CFLAGS, and let autoconf add -O0.