core: Fix multiple issues with discovery handling
This patch is the first attempt in fixing the discovery issues for
dual-mode BR/EDR + LE controllers. It creates a clean chain for
start/stop discovery client handling and also for the internal
handling of suspend, resume and restart.

However this patch removes the LE connection handling. And that needs
to be put back with a separate patch.