adapter: Fix not removing a device when the adapter is down
It is a documented error that MGMT_UNPAIR_DEVICE should return an error
if the controller is not powered.

This error is being ignored, and if the user asks for a device to be
removed while the controller is powered down, the keys associated with
that device will still be in the kernel side.

The solution is to return the user an error when it is asked to remove a
device and the controller is not powered.