core: Clear device->attrib pointer before calling g_attrib_cancel_all
The g_attrib_cancel_all function may call functions such as
find_included_cb that should at this point no longer have access to the
GAttrib instance and try to do operations with it. Therefore it is
safest to ensure that device->attrib is no longer available in these
functions.