core: Fix emitting empty Device1.UUIDs property
In case there is a incoming connection but services has not been
resolved the UUID of the profile being connected is automatically added
using btd_device_add_uuid which does update uuids not eir_uuids which is
used if svc_resolved is false.

To fix this the code will now fallback to uuids field whenever eir_uuids is
empty so uuids added with btd_device_add_uuid previous to services being
resolved should appear in UUIDs property.