obexd: Add obc_session_get_folder function
This patch extends client session by the tracking of the current folder.
The current folder can be accessed by obc_session_get_current_folder.

This allows drivers to add a folder property to browsed objects so that
the application doesn't have to keep track of the folder an object
belongs to.