audio/AVCTP: Fix sending requests with same transaction id
If a request is outstanding in the processed list its transaction shall
not be reused as it can cause the wrong callback to be called.

This can be reproduced in very rare occasions where e.g. a notification
using the same transaction of the current request arrives before the
response itself.